Switch to touchscreen

+++ SupervisorBoardCode/lib/miam_utils_arduino/Types.h
@@ -8,6 +8,6 @@
     
     #include <ArduinoEigenDense.h>
 
-    typedef Eigen::Vector2d Vector2;
-    typedef Eigen::Vector3d Vector3;
+    typedef Eigen::Vector2f Vector2;
+    typedef Eigen::Vector3f Vector3;
 #endif
